Patch 1395
・item_dbのテーブルを更新 (item_db.txt)

・PACKETVER[20181017]以降にクエストリスト用パケットを追加 (clif.c)

・[SQL] : キャストエラー修正 (sqldbs.c)

+++ src/common/sqldbs.c
@@ -334,7 +334,7 @@ bool sqldbs_stmt_simpleprepare(struct sqldbs_stmt *st, const char *query)
 	}
 	st->query = (char *)aStrdup(query);
 
-	if( mysql_stmt_prepare(st->stmt, query, strlen(query)) )
+	if( mysql_stmt_prepare(st->stmt, query, (unsigned long)strlen(query)) )
 	{
 		printf("DB server Error - %s\n  %s\n\n", mysql_stmt_error(st->stmt), query);
 		return false;
